Move CustomButtonsSettingsFragment to a new file
